--- a/mempool/src/tests/multi_node_test.rs
+++ b/mempool/src/tests/multi_node_test.rs
@@ -24,6 +24,7 @@ use aptos_network::{
     ProtocolId,
 };
 use aptos_types::{transaction::SignedTransaction, PeerId};
+use maplit::btreemap;
 use rand::{rngs::StdRng, SeedableRng};
 use std::collections::HashMap;
 use tokio::runtime::Runtime;
@@ -376,7 +377,7 @@ impl TestHarness {
                                 102400,
                                 true,
                                 false,
-                                vec![],
+                                btreemap![],
                             );
                             for txn in transactions.iter() {
                                 assert!(block.contains(txn));