cherry pick #10744 

### Description

In forge `realistic_network_tuned_for_throughput` it was found that:
* Calling pull takes > 25 ms.
* In an edge case where the pull returns 0 new txns, this could cause livelock because the pull call would be called every 25 ms.

Resolved by:
* Optimizing the pull; now on the same workload the pull takes ~8 ms.
  * Instead of using a hashmap, use a sorted vector + binary search for the excluded txns
  * And only re-sort the vector when batches in progress is updated (instead of every pull)
* Checking the time taken to pull, and pulling less frequently if the pull takes > 12.5 ms.

### Test Plan

Existing tests, ran `realistic_network_tuned_for_throughput` as ad-hoc job and observed results.

### Description

### Test Plan
<!-- Please provide us with clear details for verifying that your changes work. -->